server/requirements.txt
python-socketio
eventlet
redis